The fishes are now alive

+++ aquarium/items/Seaweed.java
@@ -31,6 +31,6 @@ public class Seaweed extends AquariumItem
 
 	@Override
 	public void draw(Graphics g) {
-		g.drawImage(image, position.x, position.y, width, height, null);
+		g.drawImage(image, (int)position.x, (int)position.y, width, height, null);
 	}
 }
